# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: EcoRecicla Bogotá API - Recycling Management System for Residential Complexes
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /records:
    get:
      operationId: listRecords
      tags: [records]
      summary: List all recycling records
      parameters:
        - name: residentId
          in: query
          schema:
            type: integer
        - name: complexId
          in: query
          schema:
            type: integer
        - name: materialId
          in: query
          schema:
            type: integer
        - name: dateFrom
          in: query
          schema:
            type: string
        - name: dateTo
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: pageSize
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of records
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecordListResponse"
    post:
      operationId: createRecord
      tags: [records]
      summary: Create a recycling record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecordInput"
      responses:
        "201":
          description: Record created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecyclingRecord"
  /records/{id}:
    get:
      operationId: getRecord
      tags: [records]
      summary: Get a record by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecyclingRecord"
        "404":
          description: Not found
    patch:
      operationId: updateRecord
      tags: [records]
      summary: Update a record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecordUpdate"
      responses:
        "200":
          description: Updated record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecyclingRecord"
        "404":
          description: Not found
    delete:
      operationId: deleteRecord
      tags: [records]
      summary: Delete a record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /records/total-kg:
    get:
      operationId: getTotalKg
      tags: [records]
      summary: Get total kilograms recycled
      parameters:
        - name: complexId
          in: query
          schema:
            type: integer
        - name: materialId
          in: query
          schema:
            type: integer
        - name: residentId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Total kg
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TotalKgResponse"

  # DASHBOARD
...
  /reports/summary:
    get:
      operationId: getReportSummary
      tags: [reports]
      summary: Get summary report data
      parameters:
        - name: dateFrom
          in: query
          schema:
            type: string
        - name: dateTo
          in: query
          schema:
            type: string
        - name: materialId
          in: query
          schema:
            type: integer
        - name: residentId
          in: query
          schema:
            type: integer
        - name: complexId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Report summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportSummary"

  # PROFILE
...
components:
  schemas:
...
    RecyclingRecord:
      type: object
      required: [id, residentId, complexId, materialId, weightKg, date, createdAt]
      properties:
        id:
          type: integer
        residentId:
          type: integer
        residentName:
          type: ["string", "null"]
        complexId:
          type: integer
        complexName:
          type: ["string", "null"]
        materialId:
          type: integer
        materialName:
          type: ["string", "null"]
        weightKg:
          type: number
        date:
          type: string
        observation:
          type: ["string", "null"]
        responsibleUserId:
          type: ["integer", "null"]
        responsibleUserName:
          type: ["string", "null"]
        createdAt:
          type: string
    RecordInput:
      type: object
      required: [residentId, complexId, materialId, weightKg, date]
      properties:
        residentId:
          type: integer
        complexId:
          type: integer
        materialId:
          type: integer
        weightKg:
          type: number
        date:
          type: string
        observation:
          type: string
        responsibleUserId:
          type: integer
    RecordUpdate:
      type: object
      properties:
        residentId:
          type: integer
        complexId:
          type: integer
        materialId:
          type: integer
        weightKg:
          type: number
        date:
          type: string
        observation:
          type: string
        responsibleUserId:
          type: integer
    RecordListResponse:
      type: object
      required: [data, total, page, pageSize, totalKg]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/RecyclingRecord"
        total:
          type: integer
        page:
          type: integer
        pageSize:
          type: integer
        totalKg:
          type: number
    TotalKgResponse:
      type: object
      required: [totalKg]
      properties:
        totalKg:
          type: number

    # DASHBOARD
...
    MaterialBreakdownItem:
      type: object
      required: [materialId, materialName, binColor, totalKg, totalRecords]
      properties:
        materialId:
          type: integer
        materialName:
          type: string
        binColor:
          type: string
        totalKg:
          type: number
        totalRecords:
          type: integer

    # REPORTS
    ReportSummary:
      type: object
      required: [totalKg, totalRecords, byMaterial, byComplex, byResident, records]
      properties:
        totalKg:
          type: number
        totalRecords:
          type: integer
        byMaterial:
          type: array
          items:
            $ref: "#/components/schemas/MaterialBreakdownItem"
        byComplex:
          type: array
          items:
            $ref: "#/components/schemas/ComplexBreakdownItem"
        byResident:
          type: array
          items:
            $ref: "#/components/schemas/ResidentBreakdownItem"
        records:
          type: array
          items:
            $ref: "#/components/schemas/RecyclingRecord"
    ComplexBreakdownItem:
      type: object
      required: [complexId, complexName, totalKg, totalRecords]
      properties:
        complexId:
          type: integer
        complexName:
          type: string
        totalKg:
          type: number
        totalRecords:
          type: integer
    ResidentBreakdownItem:
      type: object
      required: [residentId, residentName, totalKg, totalRecords]
      properties:
        residentId:
          type: integer
        residentName:
          type: string
        totalKg:
          type: number
        totalRecords:
          type: integer

    # PROFILE
...