# lib/api-zod/src/generated/types/residentBreakdownItem.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * EcoRecicla Bogotá API - Recycling Management System for Residential Complexes
 * OpenAPI spec version: 0.1.0
 */

export interface ResidentBreakdownItem {
  residentId: number;
  residentName: string;
  totalKg: number;
  totalRecords: number;
}